sales_team: extract prompts into prompts/ package (#252)
Move the 10 inline _X_SYSTEM_PROMPT constants and the matching inline
f-string user prompts out of agents.py into a per-specialist prompts/
package, mirroring the blogging team convention.

Each leaf module exports SYSTEM_PROMPT (base + rendered FEWSHOT_EXAMPLES,
which is empty in this commit so SYSTEM_PROMPT is byte-identical to the
prior inline constant) and a TASK_TEMPLATE used at the call site via
.format(**ctx). The dossier-rendering helper moves into the prompts
package alongside the outreach prompt that depends on it.

Tests in tests/test_prompts.py assert each TASK_TEMPLATE.format() output
matches the original f-string output for representative inputs, and
that every SYSTEM_PROMPT is non-empty.

agents.py drops from 1,110 to 488 lines; the issue's <400 LOC target
isn't reached because the 11-class dataclass scaffolding accounts for
~80 lines that can't compress without a base-class refactor that's
out of scope here. The four "dead JSON skeleton blocks" the issue
references are not present in the file (already cleaned up).
